update browser tests
